+++ pkg/scalers/azure/azure_queue_test.go
@@ -2,12 +2,13 @@ package azure
 
 import (
 	"context"
+	"net/http"
 	"strings"
 	"testing"
 )
 
 func TestGetQueueLength(t *testing.T) {
-	length, err := GetAzureQueueLength(context.TODO(), "", "", "queueName", "")
+	length, err := GetAzureQueueLength(context.TODO(), http.DefaultClient, "", "", "queueName", "")
 	if length != -1 {
 		t.Error("Expected length to be -1, but got", length)
 	}
@@ -20,7 +21,7 @@ func TestGetQueueLength(t *testing.T) {
 		t.Error("Expected error to contain parsing error message, but got", err.Error())
 	}
 
-	length, err = GetAzureQueueLength(context.TODO(), "", "DefaultEndpointsProtocol=https;AccountName=name;AccountKey=key==;EndpointSuffix=core.windows.net", "queueName", "")
+	length, err = GetAzureQueueLength(context.TODO(), http.DefaultClient, "", "DefaultEndpointsProtocol=https;AccountName=name;AccountKey=key==;EndpointSuffix=core.windows.net", "queueName", "")
 
 	if length != -1 {
 		t.Error("Expected length to be -1, but got", length)