+++ pkg/scalers/azure_eventhub_scaler_test.go
@@ -3,6 +3,7 @@ package scalers
 import (
 	"context"
 	"fmt"
+	"net/http"
 	"net/url"
 	"os"
 	"testing"
@@ -111,7 +112,7 @@ func TestGetUnprocessedEventCountInPartition(t *testing.T) {
 
 	if eventHubKey != "" && storageConnectionString != "" {
 		eventHubConnectionString := fmt.Sprintf("Endpoint=sb://%s.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=%s;EntityPath=%s", testEventHubNamespace, eventHubKey, testEventHubName)
-		storageCredentials, endpoint, err := azure.ParseAzureStorageBlobConnection("none", storageConnectionString, "")
+		storageCredentials, endpoint, err := azure.ParseAzureStorageBlobConnection(http.DefaultClient, "none", storageConnectionString, "")
 		if err != nil {
 			t.Error(err)
 			t.FailNow()
@@ -438,7 +439,11 @@ func TestEventHubGetMetricSpecForScaling(t *testing.T) {
 		if err != nil {
 			t.Fatal("Could not parse metadata:", err)
 		}
-		mockEventHubScaler := azureEventHubScaler{meta, nil}
+		mockEventHubScaler := azureEventHubScaler{
+			metadata:   meta,
+			client:     nil,
+			httpClient: http.DefaultClient,
+		}
 
 		metricSpec := mockEventHubScaler.GetMetricSpecForScaling()
 		metricName := metricSpec[0].External.Metric.Name